+++ K8sUtils/K8sUtils.psm1
@@ -15,7 +15,7 @@ Set-K8sUtilsConfig
 if (!$Quiet -and !(Test-Path env:TF_BUILD)) {
     $me = $MyInvocation.MyCommand.Name -split '\.' | Select-Object -First 1
     Write-Information "`n$me loaded. Use help <command> -Full for help.`n`nCommands:" -InformationAction Continue
-    $exports | Write-Information -InformationAction Continue
+    $exports | Sort-Object | Write-Information -InformationAction Continue
 
     Write-Information "`nUse Import-Module $me -ArgumentList `$true to suppress this message`n" -InformationAction Continue
 }