+++ Tools/Deploy-MinimalJob.ps1
@@ -26,6 +26,24 @@ Tag to use for the init container, defaults to latest
 .PARAMETER SkipRollbackOnError
 If set, don't do a helm rollback on error
 
+.PARAMETER TimeoutSecs
+How long to wait for the job to complete, defaults to 600 seconds
+
+.PARAMETER PollIntervalSec
+How often to poll for the job to complete, defaults to 3 seconds
+
+.PARAMETER ColorType
+Type of color to use for output, defaults to ANSI, can be None or DevOps
+
+.PARAMETER BadSecret
+If set, use a bad secret name in the job
+
+.PARAMETER PassThru
+If set, return the ReleaseStatus and PodStatus objects instead of writing to the console
+
+.PARAMETER ActiveDeadlineSeconds
+How long to wait for the job to complete, defaults to 30 seconds
+
 .EXAMPLE
 
 .OUTPUTS
@@ -49,7 +67,8 @@ function Deploy-MinimalJob {
         [ValidateSet("None", "ANSI", "DevOps")]
         [string] $ColorType = "ANSI",
         [switch] $BadSecret,
-        [switch] $PassThru
+        [switch] $PassThru,
+        [int] $ActiveDeadlineSeconds = 30
     )
     Set-StrictMode -Version Latest
     $ErrorActionPreference = "Stop"
@@ -99,10 +118,12 @@ function Deploy-MinimalJob {
                 "service.enabled=false",
                 "preHook.create=false",
                 "ingress.enabled=false",
+                "jobActiveDeadlineSeconds=$activeDeadlineSeconds",
                 "job.create=true",
                 "job.fail=$Fail",
                 "job.imageTag=$ImageTag",
                 "job.runCount=$RunCount"
+
 
     Write-Verbose ("HelmSet:`n   "+($helmSet -join "`n   "))
     $releaseName = "test"